[#13494] DB: Add Support for making /tmp directory customisable
Summary:
This diff adds support to use a custom temporary directory instead of the default `/tmp`. This is a requirement from customers who are reluctant to provide the execute access to the `/tmp` directory because of compliance and security reasons.

Yugabyte DB stores all the temporary files under the `/tmp` directory.

To support custom temporary directory, a new GFLAG `yb_tmp_dir` is introduced in this diff. The default value of the flag is empty string in which case, the default directory of `/tmp` will be used. When set with a non-empty value, the provided path will be used to store temporary files.

Test Plan:
Running all the C++ unit tests as well as running those impact scenarios.

New test cases:
`./yb_build.sh --cxx-test hash_util-test --gtest_filter HashUtilTest.PgSocketDerivationWithCustom`

Reviewers: skumar, smishra, rvenkatesh, hsunder, rskannan, jason

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D18798