[#20726] YSQL: Integrate Walsender with GetConsistentChanges RPC
Summary:
This revision integrates the LogicalReplication via the Walsender to use the GetConsistentChanges RPC from the CDC service.

This lifts the restriction on the number of tables and tablets supported. Before this revision, only a single tablet was supported for streaming.

Also, this revision removes all the tablet and checkpoint related changes from the Walsender as all that is handled inside the VirtualWAL component present in the CDC
service.

**Upgrade/Rollback safety:**
These changes are disabled using the test flag: ysql_TEST_enable_replication_slot_consumption

Test Plan: ```./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot'```

Differential Revision: https://phorge.dev.yugabyte.com/D32673