[#3109][#26811][#26805] YSQL: Fix separate ddl transaction condition for CREATE INDEX
Summary:
With the unified DDL/DML transaction support, only CREATE INDEX can run as a separate DDL transaction. Specifically, it runs in the following condition:
1. No explicit transaction block
2. CREATE INDEX or CREATE INDEX CONCURRENTLY

The determination of these conditions is done in GetDdlMode function in pg_yb_utils.c and before this change, it relied on the condition `stmt->concurrent !=
YB_CONCURRENCY_DISABLED`. This condition is incomplete, because the stmt->concurrent value can be YB_CONCURRENCY_IMPLICIT_ENABLED even inside a transaction block. We
only change the value to `YB_CONCURRENCY_DISABLED` at a later stage of execution (ProcessUtilitySlow).

The fix is to also check for explicit transaction block as a condition.

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestDdlTransactionBlocks'

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43199