[#29143,#29148] YSQL: Fix TestPgRegressIndex with and without txn ddl
Summary:
The test TestPgRegressIndex fails with transactional ddl. It fails because of change in how REINDEX INDEX works for partitioned tables with
transactional ddl. When transactional DDL is enabled, we follow the PG semantics of starting one transaction for each partition. So, if there is an
error during REINDEX of third partition, the reindex of the first 2 partitions is still complete and successful. This was not the case without txn ddl
because the separate DDL transaction was at the statement level and upon any error, the entire DDL transaction spanning across the partition would
rollback.

The fix is to simply adjust the sql statement according to the new behavior. This revision also duplicates the test for the case of without txn ddl.
This is because we wish to run both variants for at least 1-2 more releases before Txn DDL becomes an auto-flag.

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressIndex#schedule'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressIndex#scheduleWithoutTxnDdl'

Differential Revision: https://phorge.dev.yugabyte.com/D48263