[#29450] YSQL: Retry DDLs due to conflict with RC isolation when DDL savepoint is enabled
Summary:
As part of transactional DDL support with READ COMMITTED isolation level, we disallowed retrying DDL statements due to conflicts. This was
done as part of D44939 / 4d472acc5595. This was done because we lacked the support for rolling back DDL statements via savepoints.

This revision now allows the retry of DDLs when DDL savepoint support is enabled.

Also fixed two issues with DDL retries:
1. Cache invalidation in case of sub-transaction rollback in AtEOSubXact_Inval. We also take into account the invalidation messages from the current
statement due to aggressive caching used by YSQL. This is similar to how it is done during transaction rollback in AtEOXact_Inval.

2. Not clearing DDL state when recreating the transaction in pg_txn_manager

Test Plan:
Enabled transactional DDL and ran
./yb_build.sh --cxx-test pgwrapper_pg_wait_on_conflict-test --gtest_filter PgWaitQueuesTestWithoutObjectLocking.TestDDLHaveWaitStartTimeSet

The test fails without the fix (3 out of 5 times) and passes all 5 times.

Differential Revision: https://phorge.dev.yugabyte.com/D49145